rag_gold/requirements.txt
faiss-cpu
llama-index-readers-file
llama-index-embeddings-huggingface
llama-index-llms-groq
llama-index-llms-anthropic
llama-index-vector-stores-faiss
llama-index
pymupdf
